+++ src/dso/exec.py
@@ -1,6 +1,8 @@
 import os
+import stat
 import subprocess
 import sys
+import tempfile
 from contextlib import contextmanager
 from pathlib import Path
 from textwrap import dedent, indent
@@ -29,7 +31,35 @@ def _render_quarto(quarto_dir: Path, report_dir: Path, before_script: str, cwd: 
     before_script = indent(before_script, " " * 8)
     report_dir = report_dir.absolute()
     report_dir.mkdir(exist_ok=True)
-    pandocfilter = "--filter dso_pandocfilter" if with_pandocfilter else ""
+
+    # clean up existing `.rmarkdown` files that may interfere with rendering
+    # these are leftovers from a previous, failed `quarto render` attempt. If they still exist, the next attempt
+    # fails. We remove them *before* the run instead of cleaning them up *after* the run, because they
+    # may be usefule for debugging failures.
+    # see https://github.com/Boehringer-Ingelheim/dso/issues/54
+    for f in quarto_dir.glob("*.rmarkdown"):
+        if f.is_file():
+            f.unlink()
+
+    # Enable pandocfilter if requested.
+    # We create a temporary script that then calls the current python binary with the dso.pandocfilter module
+    # This may seem cumbersome, but we do it this way because
+    #  * pandoc only supports a single binary for `--filter`, referring to subcommands or `-m` is not possible here
+    #  * we want to ensure that exactly the same python/dso version is used for the pandocfilter as for the
+    #    parent command (important when running through dso-mgr)
+    filter_script = None
+    if with_pandocfilter:
+        with tempfile.NamedTemporaryFile(delete=False, mode="w") as f:
+            f.write("#!/bin/bash\n")
+            f.write(f'{sys.executable} -m dso.pandocfilter "$@"\n')
+            filter_script = Path(f.name)
+
+        filter_script.chmod(filter_script.stat().st_mode | stat.S_IEXEC)
+
+        pandocfilter = f"--filter {filter_script}"
+    else:
+        pandocfilter = ""
+
     # propagate quiet setting to quarto
     quiet = "--quiet" if bool(int(os.environ.get("DSO_QUIET", 0))) else ""
     script = dedent(
@@ -46,6 +76,11 @@ def _render_quarto(quarto_dir: Path, report_dir: Path, before_script: str, cwd: 
         """
     )
     res = subprocess.run(script, shell=True, executable="/bin/bash", cwd=cwd)
+
+    # clean up
+    if filter_script is not None:
+        filter_script.unlink()
+
     if res.returncode:
         sys.exit(res.returncode)